refactor: RedisModule 위치 변경 issue: SCRUM-118
