+++ src/app.module.ts
@@ -6,11 +6,11 @@ import { BookmarkModule } from './modules/bookmarks/bookmark.module';
 import { TeamModule } from './modules/teams/team.module';
 import { UserModule } from './modules/users/user.module';
 import { ResumeModule } from './modules/resumes/resume.module';
-import { RedisModule } from './auth/auth.redis.module';
 import { AuthModule } from './auth/auth.module';
 import { SessionModule } from './modules/sessions/session.module';
 import { EventModule } from './modules/events/event.module';
 import { LikeModule } from './modules/likes/like.module';
+import { RedisModule } from './global/redis/redis.module';
 
 @Module({
     imports: [
