+++ src/auth/auth.module.ts
@@ -3,8 +3,8 @@ import { JwtModule } from '@nestjs/jwt';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { JwtStrategy } from './jwt.strategy';
-import { RedisModule } from './auth.redis.module';
 import { UserModule } from '../modules/users/user.module'; // UserModule과의 순환참조 방지
+import { RedisModule } from '../global/redis/redis.module';
 
 @Module({
     imports: [
