+++ src/global/redis/redis.module.ts
@@ -1,7 +1,7 @@
 import { Module, Global } from '@nestjs/common';
 import Redis from 'ioredis';
 import { ConfigModule } from '@nestjs/config';
-import { PrismaService } from '../modules/prisma/prisma.service';
+import { PrismaService } from '../../modules/prisma/prisma.service';
 
 @Global() // 전역 모듈로 설정하여 애플리케이션 어디서나 사용 가능
 @Module({
